DASHBOARD/requirement.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
click==6.7
dash-core-components==0.3.3
dash-html-components==0.12.0
dash-renderer==0.2.9
dash.ly==0.14.0
decorator==4.0.11
enum34==1.1.6
Flask==0.12
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
funcsigs==1.0.2
functools32==3.2.3.post2
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.6.0
jupyter-core==4.3.0
MarkupSafe==1.0
mock==2.0.0
nbformat==4.3.0
numpy==1.11.0
pathlib2==2.2.1
pbr==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
plotly==2.0.8
pluggy==0.4.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.33
Pygments==2.2.0
python-dateutil==2.5.3
pytz==2017.2
requests==2.13.0
requests-file==1.4.1
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
tox==2.7.0
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.1